app/requirements.txt
numpy==1.26.3
streamlit==1.30.0
tensorflow==2.15.0.post1
flask==2.3.2
flask-cors==3.0.10
Pillow
boto3
python-dotenv
